+++ src/Kiota.Builder/EqualityComparers/OpenApiSchemaReferenceComparer.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using Microsoft.OpenApi.Models;
+
+namespace Kiota.Builder.EqualityComparers;
+
+internal class OpenApiSchemaReferenceComparer : IEqualityComparer<OpenApiSchema>
+{
+    public bool Equals(OpenApiSchema? x, OpenApiSchema? y)
+    {
+        return (x, y) switch
+        {
+            (null, null) => true,
+            (null, _) => false,
+            (_, null) => false,
+            _ when x.Reference is not null && !string.IsNullOrEmpty(x.Reference.Id) && y.Reference is not null && !string.IsNullOrEmpty(y.Reference.Id) =>
+                    x.Reference.Id.Equals(y.Reference.Id, StringComparison.OrdinalIgnoreCase),
+            _ => x == y,
+        };
+    }
+    public int GetHashCode([DisallowNull] OpenApiSchema obj)
+    {
+        return obj.Reference is not null && !string.IsNullOrEmpty(obj.Reference.Id) ? obj.Reference.Id.GetHashCode(StringComparison.OrdinalIgnoreCase) : obj.GetHashCode();
+    }
+}