+++ src/Kiota.Builder/Extensions/OpenApiOperationExtensions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-
+using Kiota.Builder.Configuration;
 using Microsoft.OpenApi.Models;
 
 namespace Kiota.Builder.Extensions;
@@ -13,33 +13,33 @@ public static class OpenApiOperationExtensions
     /// cleans application/vnd.github.mercy-preview+json to application/json
     /// </summary>
     private static readonly Regex vendorSpecificCleanup = new(@"[^/]+\+", RegexOptions.Compiled, Constants.DefaultRegexTimeout);
-    public static OpenApiSchema? GetResponseSchema(this OpenApiOperation operation, HashSet<string> structuredMimeTypes)
+    internal static OpenApiSchema? GetResponseSchema(this OpenApiOperation operation, StructuredMimeTypesCollection structuredMimeTypes)
     {
         ArgumentNullException.ThrowIfNull(operation);
         // Return Schema that represents all the possible success responses!
         return operation.GetResponseSchemas(SuccessCodes, structuredMimeTypes)
                             .FirstOrDefault();
     }
-    internal static IEnumerable<OpenApiSchema> GetResponseSchemas(this OpenApiOperation operation, HashSet<string> successCodesToUse, HashSet<string> structuredMimeTypes)
+    internal static IEnumerable<OpenApiSchema> GetResponseSchemas(this OpenApiOperation operation, HashSet<string> successCodesToUse, StructuredMimeTypesCollection structuredMimeTypes)
     {
         // Return Schema that represents all the possible success responses!
         return operation.Responses.Where(r => successCodesToUse.Contains(r.Key))
                             .OrderBy(static x => x.Key, StringComparer.OrdinalIgnoreCase)
                             .SelectMany(re => re.Value.Content.GetValidSchemas(structuredMimeTypes));
     }
-    public static OpenApiSchema? GetRequestSchema(this OpenApiOperation operation, HashSet<string> structuredMimeTypes)
+    internal static OpenApiSchema? GetRequestSchema(this OpenApiOperation operation, StructuredMimeTypesCollection structuredMimeTypes)
     {
         ArgumentNullException.ThrowIfNull(operation);
         return operation.RequestBody?.Content
                             .GetValidSchemas(structuredMimeTypes).FirstOrDefault();
     }
-    private static readonly HashSet<string> multipartMimeTypes = new(1, StringComparer.OrdinalIgnoreCase) { "multipart/form-data" };
-    public static bool IsMultipartFormDataSchema(this IDictionary<string, OpenApiMediaType> source, HashSet<string> structuredMimeTypes)
+    private static readonly StructuredMimeTypesCollection multipartMimeTypes = new(new string[] { "multipart/form-data" });
+    internal static bool IsMultipartFormDataSchema(this IDictionary<string, OpenApiMediaType> source, StructuredMimeTypesCollection structuredMimeTypes)
     {
         return source.GetValidSchemas(structuredMimeTypes).FirstOrDefault() is OpenApiSchema schema &&
         source.GetValidSchemas(multipartMimeTypes).FirstOrDefault() == schema;
     }
-    public static IEnumerable<OpenApiSchema> GetValidSchemas(this IDictionary<string, OpenApiMediaType> source, HashSet<string> structuredMimeTypes)
+    internal static IEnumerable<OpenApiSchema> GetValidSchemas(this IDictionary<string, OpenApiMediaType> source, StructuredMimeTypesCollection structuredMimeTypes)
     {
         if (!(structuredMimeTypes?.Any() ?? false))
             throw new ArgumentNullException(nameof(structuredMimeTypes));
@@ -51,7 +51,7 @@ public static class OpenApiOperationExtensions
                             .Where(static s => s is not null) ??
                         Enumerable.Empty<OpenApiSchema>();
     }
-    public static OpenApiSchema? GetResponseSchema(this OpenApiResponse response, HashSet<string> structuredMimeTypes)
+    internal static OpenApiSchema? GetResponseSchema(this OpenApiResponse response, StructuredMimeTypesCollection structuredMimeTypes)
     {
         ArgumentNullException.ThrowIfNull(response);
         return response.Content.GetValidSchemas(structuredMimeTypes).FirstOrDefault();