+++ tests/Kiota.Builder.Tests/Writers/Go/CodeMethodWriterTests.cs
@@ -1293,6 +1293,56 @@ public class CodeMethodWriterTests : IDisposable
         AssertExtensions.CurlyBracesAreClosed(result);
     }
     [Fact]
+    public void WritesRequestGeneratorBodyKnownRequestBodyType()
+    {
+        setup();
+        method.Kind = CodeMethodKind.RequestGenerator;
+        method.HttpMethod = HttpMethod.Post;
+        AddRequestProperties();
+        AddRequestBodyParameters(useComplexTypeForBody: false);
+        method.Parameters.OfKind(CodeParameterKind.RequestBody).Type = new CodeType
+        {
+            Name = new GoConventionService().StreamTypeName,
+            IsExternal = true,
+        };
+        method.RequestBodyContentType = "application/json";
+        writer.Write(method);
+        var result = tw.ToString();
+        Assert.Contains("SetStreamContentAndContentType", result, StringComparison.OrdinalIgnoreCase);
+        Assert.Contains("application/json", result, StringComparison.OrdinalIgnoreCase);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
+    public void WritesRequestGeneratorBodyUnknownRequestBodyType()
+    {
+        setup();
+        method.Kind = CodeMethodKind.RequestGenerator;
+        method.HttpMethod = HttpMethod.Post;
+        AddRequestProperties();
+        AddRequestBodyParameters(useComplexTypeForBody: false);
+        method.Parameters.OfKind(CodeParameterKind.RequestBody).Type = new CodeType
+        {
+            Name = new GoConventionService().StreamTypeName,
+            IsExternal = true,
+        };
+        method.AddParameter(new CodeParameter
+        {
+            Name = "requestContentType",
+            Type = new CodeType()
+            {
+                Name = "string",
+                IsExternal = true,
+            },
+            Kind = CodeParameterKind.RequestBodyContentType,
+        });
+        writer.Write(method);
+        var result = tw.ToString();
+        Assert.Contains("SetStreamContentAndContentType", result, StringComparison.OrdinalIgnoreCase);
+        Assert.DoesNotContain("application/json", result, StringComparison.OrdinalIgnoreCase);
+        Assert.Contains(", requestContentType", result, StringComparison.OrdinalIgnoreCase);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
     public void WritesInheritedDeSerializerBody()
     {
         setup(true);