+++ tests/Kiota.Builder.Tests/Writers/TypeScript/CodeFileWriterTests.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Kiota.Builder.CodeDOM;
 using Kiota.Builder.Configuration;
 using Kiota.Builder.Extensions;
+using Kiota.Builder.OrderComparers;
 using Kiota.Builder.Refiners;
 using Kiota.Builder.Writers;
 using Xunit;